[core] Converted Media RSS structs to GObjects
Converted all the Media RSS structs to GObjects, and moved them to a "media"
directory. The documentation for the new objects is mostly complete, but no
new test cases have been added.
A few of the new classes had to be subclassed to add YouTube-specific
properties to them. Helps: bgo#579170