[build] Reorganised to use non-recursive automake and added gcov support
Reorganised to use non-recursive automake for everything except API
documentation and translations. As a consequence of this reorganisation, two
private header files which were mistakenly installed before are no longer
installed (this will not be considered an API break). The introspection and
gtester include files have been updated.

Support for code coverage checking for the test suite has been added too,
using gcov. It can be invoked by `make gcov`, and will output a file called
"gcov-report.txt" in the main source directory with the results of the
checks.