Bug 593239 — Use GDataUploadStream for uploading files
Use GDataUploadStream in the PicasaWeb service to upload photos. This also
now adds the appropriate kind category to entries when they're uploaded (if
it hasn't already been added to the entry), and improves the upload test case
so that the current timestamp is added to the description for uploaded files,
allowing for easier inspection of files online to see whether the tests
were successful. Closes: bgo#593239