[core] Remove several enumerated types
Remove GDataYouTubeAspectRatio, and replace it with a
GDATA_YOUTUBE_ASPECT_RATIO_WIDESCREEN #define.

Remove GDataYouTubeAction and replace it with #defines for all its members.

This is an API and ABI break necessary to enforce the newly-finalised
policy that enums are avoided where possible for API which interacts with or
represents data from the server.