Bug 614253 — GDataMediaGroup:tags should be a string array
Convert GDataMediaGroup:tags to be a string array, which automatically
escapes and unescapes the tags passed to it. Also convert all properties
which proxy GDataMediaGroup:tags, which means the following API breaks:
 * GDataPicasaWebAlbum:tags and getter/setter
 * GDataPicasaWebFile:tags and getter/setter
 * GDataYouTubeVideo:keywords and getter/setter
All these properties are now string arrays, and their getters/setters deal
with them as such. Closes: bgo#614253