[core] Unify queries for single entries
Add a new gdata_service_query_single_entry() function (and async variant)
to serve as the main function to call for retrieving a single entry from
any service.

As a consequence of this, the GDataQuery:entry-id property has been removed,
as it's now redundant (and never worked well anyway).

ABI has been broken apart from the API removals due to adding a new member
to GDataEntryClass.

Note that this includes the removal of:
 * gdata_youtube_service_query_single_video() (and async variant)
 * gdata_documents_service_query_single_document()
 * gdata_query_new_for_id()
 * gdata_query_[get|set]_entry_id()