Bug 635959 — Proper cancellation support for GDataUploadStream
Add proper cancellation support to GDataUploadStream. This involves adding
support for not starting the upload thread if the operation has been
cancelled very early, plus adding support for cancelling it mid-upload using
soup_session_cancel_message().

One of the PicasaWeb upload cancellation tests had to be changed to use a
smaller timeout in order to provide more of a guarantee that it'll cancel
during the upload.

Closes: bgo#635959