Bug 635736 — Asynchronous authentication tests broken
The asynchronous authentication tests have been broken for a while, probably
caused due to changes of asynchronous callback functions within GIO. We
previously depended on an idle function to set the username and password
after a successful asynchronous authentication, but due to the priority
changes this has stopped being called before the asynchronous result
callback, meaning that the results of authentication weren't available to
the callback. This made it look like authentication had failed, when in
reality it had succeeded and the results were waiting in the main loop's
queue.

This changes it so that the authentication information in GDataService is
set inside the authentication thread, protected by a mutex.

Closes: bgo#635736