documents: Work around an ETag problem when uploading new documents
Google's servers appear to modify newly uploaded documents a few seconds
after returning the entry and ETag for them to the client, thus changing the
documents' ETags and causing subsequent queries by the client to fail due to
an outdated ETag. This commit hacks the test suite to sleep for 5 seconds
after uploading a new document before querying for the new document to get
the updated ETag.

See:
http://code.google.com/a/google.com/p/apps-api-issues/issues/detail?id=2337