core: Emit notifications from GDataDownloadStream in the download thread
Emitting them in the main thread doesn't guarantee us any extra thread
safety (the client is still required to be MT-safe, as the thread which
started the download operation could be different from the one with the
main loop), and can cause memory leaks if a main loop isn't running.

Note that this is an API break (of sorts), as extra restrictions have been
added to the use of notification signals for
GDataDownloadStream::content-length and GDataDownloadStream::content-type.