picasaweb: Switch to a stream-based upload API
For similar reasons as the Documents service's switch to a stream-based
upload and download API, it makes sense to switch the PicasaWeb service to a
stream-based upload API. This gives us more flexibility as to the source of
uploaded data.

The test suite has been updated appropriately.

The following API has been changed:
 • gdata_picasaweb_service_upload_file()

The following API has been removed:
 • gdata_picasaweb_service_upload_file_async()
 • gdata_picasaweb_service_upload_file_finish()

The following API has been added:
 • gdata_picasaweb_service_finish_file_upload()