media: Switch GDataMediaThumbnail to a stream-based download API
Downloading to a file provided by its path was unnecessarily limiting,
forcing downloads to be made directly to disk instead of (e.g.) another
network stream, and subjecting clients to libgdata's idea of “Replace this
File?” logic. It's not much extra code for the clients to handle this
themselves, and offers much greater flexibility. We also get async support
for free, as creating the GDataDownloadStream is a cheap operation, and it
already has async support.

This updates the test suite appropriately. We now load thumbnails in the
test suite directly into GdkPixbufs, which requires that we bump our
gdk-pixbuf requirement to 2.14.

This is an API break which affects:
 • gdata_media_thumbnail_download()

Helps: bgo#633363