youtube: Switch to a stream-based upload API
For similar reasons as the Documents service's switch to a stream-based
upload and download API, it makes sense to switch the YouTube service to a
stream-based upload API. This gives us more flexibility as to the source of
uploaded data.

The test suite has been updated appropriately.

The following API has been changed:
 • gdata_youtube_service_upload_video()

The following API has been removed:
 • gdata_youtube_service_upload_video_async()
 • gdata_youtube_service_upload_video_finish()

The following API has been added:
 • gdata_youtube_service_finish_video_upload()