documents: Disconnect from the content-type notification signal
We can't change GDataDownloadStream to emit notification signals for
content-type in the download thread in the 0.7 branch, since that would be
an API break. This means that the notifications could potentially be
delivered after gdata_documents_document_download() returns. Since the
closure for the notification idle function maintains a reference to the
GDataDownloadStream, this means that the signal handler won't be disconnected
automatically once the GDataDownloadStream is unreffed in
gdata_documents_document_download(). Consequently, it's possible to get
invalid writes into random bits of memory unless we disconnect from the
content-type notification signal. This commit does that.