documents: Disconnect from the content-type notification signal
We can't change GDataDownloadStream to emit notification signals for
content-type in the download thread in the 0.6 branch, since that would be
an API break. This means that the notifications could potentially be
delivered after _gdata_documents_entry_download_document() returns. Since the
closure for the notification idle function maintains a reference to the
GDataDownloadStream, this means that the signal handler won't be disconnected
automatically once the GDataDownloadStream is unreffed in
_gdata_documents_entry_download_document(). Consequently, it's possible to get
invalid writes into random bits of memory unless we disconnect from the
content-type notification signal. This commit does that.