core: Fix a race condition in message cancellation handling
Message cancellation is performed using soup_session_cancel_message(), but
this has no effect until after the message is queued in the SoupSession's
message queue. Consequently, with the naïve cancellation implementation we
had before, there was a race condition which could occur if the message was
cancelled before it was queued in the session (typically if the cancellable
was cancelled before the message was even passed to the GDataService).

We fix this by locking message cancellation while the message is added to
the session's message queue, and separately handling cancellation which
occurs before the message is passed to the GDataService.