core: Fix handling early cancellation of batch operations
If we left cancellation handling to the GSimpleAsyncResult, our code to
propagate errors to all the operation's callbacks was not being called if
the GCancellable was cancelled really early in running the operation (e.g.
before calling gdata_batch_operation_run[_async]()).

This commit fixes the problem by disabling GSimpleAsyncResult's cancellation
handling feature for batch operations.