core: Ensure all GDataUploadStream ops are finished before freeing the message
It was previously possible for the SoupMessage being used in an ongoing
upload to be unreffed (by dispose()) and destroyed before the ongoing network
operation finished, causing a crash. We now ensure that all network operations
are finished at the start of dispose(), rather than at the start of
finalize().