core: Fix liveness issues and cancellation problems in GDataBuffer
There was a race condition with cancellation of a gdata_buffer_pop_data()
operation where if the operation was cancelled before the first g_cond_wait()
call, the thread would sit in the g_cond_wait() for ever, as the GCond would
never be signalled in the cancellation handler, as it had already run.

This commit fixes the problem by serialising access to the cancelled flag,
ensuring that deadlock doesn't occur by setting up and tearing down the
cancellation stuff without the buffer's mutex being held.

Helps: bgo#637036