core: Fix cancellation handling for closing a GDataUploadStream
Deadlocks could have occurred if the cancellable was in the process of being
cancelled when g_cancellable_disconnect() was called (as it was being called
with the same mutex locked that the cancellation callback attempts to lock).

Additionally, the suboptimal method of running the cancellation code in an
idle callback to ensure that it wasn't run in the same thread as
gdata_upload_stream_close() can be removed, because the cancellation handler
is now installed before gdata_upload_stream_close() locks the response_mutex,
thus removing the possibility of deadlock if the cancellable has already been
cancelled.

Helps: bgo#637036