core: Don't signal completion of a download in a callback function
If completion of a GDataDownloadStream download is signalled in a callback
function, it's possible for the stream object to be unreffed and finalised
while the network thread is still using the object (as it has some cleanup
work to do after emitting the “finished” signal).

This is fixed by moving signalling of finished_cond to the end of the thread's
main function.