core: Don't assume that messages can't be cancelled externally
In _gdata_service_actually_send_message() we were previously assuming that
we were the only code which could cancel an in-progress message's request.

This isn't true.

soup_session_abort() can be called from anywhere else and will explicitly
cancel all in-progress requests. This could happen if the GDataService is
disposed of, which would cause the SoupSession to be disposed of, which
causes it to abort all in-progress requests. Alternatively, it could happen
if the GDataService's (and hence SoupSession's) proxy URI is changed.

To fix this, we:
 • now hold a reference to the session and message in
   _gdata_service_actually_send_message() so that they're kept alive
   throughout the request; and we
 • no longer assume that a cancelled message was cancelled by us, removing
   the assertion which caused the crash in bgo#650835.

Fixes: bgo#650835