Bug 653535 — Let GDataAuthorizer re-process request after refreshing
After receiving a 401 status (unauthorised) from the server and refreshing
the current GDataAuthorizer, re-process the message in question with the
authoriser before re-sending it, so that it has updated credentials set on
it.

This involves a slight change in the semantics of
GDataAuthorizer->process_request, which now has to handle being called
multiple times on a given message.

Closes: bgo#653535