documents: Fix the IDs used to refer to all documents
As explained in a comment in the commit, it turns out that the server returns
one of two different formats of ID for a document depending on how it was
created. One of these is useful, the other one is not so useful, especially
when it comes to deleting the document.