core: Add a GDataOAuth2Authorizer for OAuth 2.0 support
This adds a new GDataAuthorizer subclass to support OAuth 2.0
authorisation, which is needed for the Google Tasks unit tests.

This includes support for all Google OAuth 2.0 features (note that
Google’s implementation of OAuth 2.0 is not entirely standard, so this
OAuth 2.0 authoriser cannot be used outside of Google’s services). It
includes full unit tests too.

New API:
 • GDATA_OAUTH2_REDIRECT_URI_OOB
 • GDATA_OAUTH2_REDIRECT_URI_OOB_AUTO
 • GDataOAuth2Authorizer

https://bugzilla.gnome.org/show_bug.cgi?id=646285