core: Ensure passwords and auth. tokens are stored in non-pageable memory
If we're compiled with --enable-gnome, we now allocate passwords and auth.
tokens in non-pageable memory where possible. Where not possible (due to API
constraints imposed by ourselves or libsoup) we ensure that the details are
zeroed out before being freed.

This is all with the aim of never having passwords or auth. tokens hit disk
or be leaked in other ways.

It hasn't been formally reviewed or certified, and there are probably cases
I've missed (which are bugs).

This adds an optional dependency on libgnome-keyring when compiled with
--enable-gnome.

Helps: bgo#656783