documents: Add support for resumable uploads and updates
This adds support for resumable uploads and updates to the Documents API. In
doing so, it was necessary to wade through a sea of bugs and documentation
inadequacies in the GData protocol (thanks, Google). In the end, it all seems
to work (ignoring the pile of hacks to handle differences between
non-resumable and resumable uploads’ handling of metadata; and the bugs in
content-only uploads).

It’s worth noting that this is only a basic implementation of resumable
upload/update. It does not, for example, automatically retry chunks of
uploads which fail. That can come later.

A new, expanded set of test cases for uploads/updates is included.

New API:
 • gdata_documents_service_upload_document_resumable()
 • gdata_documents_service_update_document_resumable()

Helps: https://bugzilla.gnome.org/show_bug.cgi?id=607272
Helps: https://bugzilla.gnome.org/show_bug.cgi?id=593537