Bug 656976 — Merge eds and GNOME Documents GOA authorisers
Copy the GOA authoriser from EDS into libgdata, modifying it a bit to add
support for non-Contacts services and remove use of the liboauth HMAC function
in favour of GLib’s implementation.

This is essentially the same code as used by GNOME Documents. Originally
written by Matthew Barnes. He’s given permission for it to be relicenced from
LGPL v2 or v3 to LGPL v2.1+. (See the bug report.)

This adds the following API:
 • GDataGoaAuthorizer

This is currently untested; waiting to port EDS and GNOME Documents to it
to test it.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=656976