Bug 656970 — Add thumbnail support
Add support for downloading thumbnails of Google Documents documents. This
includes a test case, but it isn’t very useful because I can’t find a single
document which has a thumbnail expressed in its metadata. Your mileage may
vary.

This also includes an example of how to download thumbnails, but that’s
similarly (unfortunately) untestable. This leaves me with a sour taste in
my mouth.

New API:
 • gdata_documents_document_get_download_uri()

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=656970