core: Fix idle callback priority for batch operation callbacks
Occasionally, the callback for an async batch operation as a whole was being
called before the callbacks for the individual operations in the batch. This
was causing the YouTube tests to spuriously fail every so often.

Fix this by setting the callback priority for operation callbacks to be the
same as that of the entire batch.