tests: Add mock server trace support
This adds a GDataMockServer which online tests can be run against,
either to log a trace file of all HTTP messages during the test,
or to re-run the test against such a trace file on a local
(in-process) server.

Tests are transparently redirected to the mock server by overriding
the default DNS resolver, which will resolve all URIs to a loopback
interface.

This also ports GDataService from using SoupSessionSync to the new
SoupSession. SoupSession’s default HTTP timeout is 60s, whereas the
old value used by SoupSessionSync was 0s (i.e. no timeout). In order
to preserve defaults for GDataService:timeout, we need to explicitly
set the timeout to 0s.