calendar: Document mutual exclusion between event times and recurrences
Turns out the server explicitly disallows setting individual times for
an event if it has a recurrence rule set (and vice-versa). It’s not
possible to add checks for this in libgdata without breaking API, so
rely on the server to fail to insert the offending GDataCalendarEvent
if it pleases.

Spotted by Miao Yu <will.yu@aol.com>.