core: Load extra attributes from ‘kind’ categories
When a GDataEntry is created, it’s automatically given a ‘kind’ category
element, as this is the way the GData protocol identifies the schema
which applies to a given <entry>.

Previously, the ‘kind’ category from parsed XML would not override the
automatically-added category. This meant that if the XML contained extra
attributes, such as a label, they were effectively lost.

Fix this by always preferring the ‘kind’ category from the XML. Add a
test case, because test cases are cool.

https://bugzilla.gnome.org/show_bug.cgi?id=707477