Add Freebase service
This service is, according to the main site, a "A community-curated
database of well-known people, places, and things", it allows searching
for and offering information about a wide range of topics, in a
well-structured and uniform manner.

The most low-level API is the MQL query interface, that is a JSON-based
language, queries consist of a data graph (according to their data schema)
with blank places, that will be filled in in the reply.

https://bugzilla.gnome.org/show_bug.cgi?id=726486