build: Remove .travis.yml
Having spent the best part of a day iterating (in a branch) to get the
script working, I have decided it is just not worth the time.
gnome-continuous is much more suited to libgdata, and is a lot easier to
use.

Remove the .travis.yml file so nobody is tempted to believe it is
supported or works. If they ever upgrade from Ubuntu 12.04 and improve
the iteration times for setting up .travis.yml, it might be worth
another try.