core: Fix a data corruption bug in GDataBuffer
In some situations, the buffer was reading off the end of a chunk and
into undefined memory. Spotted by asan.

e.g. Take a GDataBuffer with two chunks:
 • Chunk 1, length 8192
 • Chunk 2, length 699
and with head_read_offset = 8187 from a previous read. There are thus
704 bytes unread in the buffer.

If a read of 8192 bytes is performed, 5 bytes should come from chunk 1
and the remaining 699 from chunk 2. length_remaining was being
(correctly) set to 704, but then the while loop was not being entered,
as 704 is not greater than or equal to 8192 (the length of chunk 1). The
code was then falling into the G_LIKELY case below, and attempting to
read 704 bytes from offset 8187 of chunk 1 — overreading by 699 bytes
off the end of the chunk, and never getting the 699 bytes from chunk 2.

This can be fixed by correctly taking the head_read_offset into account
for the while condition.