core: Add gzip encoding support
libsoup already transparently supported gzip encoding of data from the
server, using SoupContentDecoder, which is present by default in a
SoupSession. We were correctly sending the Accept-Encoding header, but
Google also require that the string ‘gzip’ be present in the User-Agent
header to enable gzip support.

Add a User-Agent header which supports this.

https://bugzilla.gnome.org/show_bug.cgi?id=666623