tasks: Deprecate the ‘q’ parameter of gdata_tasks_query_new()
It was never a valid parameter, and we really should have caught this
before the API was finalised. Deprecate the parameter and document that
it should only ever have NULL passed to it. Don’t use the parameter
internally.

This is necessary because Google don’t support the ‘q’ parameter for
Tasks queries — there is no way to search within task titles or notes.