tests: Add a dummy GDataAuthorizer implementation
This will come in useful when testing services which don’t support
ClientLogin authorisation, such as Google Tasks. Such services require
OAuth authorisation, which requires user interaction. That puts up a big
barrier to running the unit tests frequently, and Google seem to provide
no ‘playground’ API key which could be used to bypass this requirement
for testing purposes.

The idea is that traces are recorded for tests using an OAuth
authoriser (requiring user interaction), then the dummy authoriser is
substituted when the tests are run offline against the pre-recorded
traces. This will require custom message matching in uhttpmock to ignore
the Authorization header, which will differ between the OAuth and dummy
authorisers.