core: Virtualise feed parsing in GDataService
Split the bulk of feed parsing out into a virtual method on
GDataService. This will allow subclasses of GDataService to implement
service-specific tweaks to feed parsing, which is required for
pagination handling with Google Documents.

New API:
 • GDataServiceClass.parse_feed

This does not break ABI — it consumes one of the struct’s expansion
slots.

https://bugzilla.gnome.org/show_bug.cgi?id=741345