core: Explicitly support a final page in GDataQuery pagination
The pagination model previously used by GDataQuery was to assume that
the server would return an empty GDataFeed once the final page was
passed by calling gdata_query_next_page().

Unfortunately, the Google Documents servers don’t do that — instead,
they provide a next page URI in each GDataFeed, and don’t provide one on
the final page. GDataQuery did not support this kind of explicit
knowledge about the current page being the final one.

Add support for it with a delightful set of hacky internal functions.
Return a dummy empty feed from GDataService if passed a GDataQuery which
is flagged as having reached the end of its pagination.

https://bugzilla.gnome.org/show_bug.cgi?id=741345