core: Get the content type from the response instead of hard coding it
There are places where we need to hard code the content type used by
a specific service API. eg., when we want to upload and need to know
whether to send JSON or XML. However, in this case, we have the
response from the server and we can find the content type from that.
We are already doing it in a few other places, so let's do it here too.

https://bugzilla.gnome.org/show_bug.cgi?id=684920