youtube: Allow location to be unset in GDataYouTubeVideo
Previously, whether the location was set was controlled by another
property, but that’s gone now. Relax the bounds on the latitude and
longitude properties, and treat all out-of-bounds values as an unset
location.

This fixes the problem of a new GDataYouTubeVideo saying its location is
(0.0, 0.0) (which is a valid location).