youtube: Remove processingDetails part from single video query URI
We can only query the processingDetails if we are the video owner, which
basically means we can only (reliably) do it in the upload code, after
uploading the video. Attempting to retrieve the processingDetails for
any other video (we don’t own) results in an access denied error
response from the server.