calendar: Set the self link for each GDataCalendarEvent
Previously, this was partially solved in GDataCalendarService by
constructing the link after each query or insertion. However, this could
not work well for async operations. Instead, it appears to be possible
to build the self link from within GDataCalendarEvent, as the calendar
ID is apparently available as the organiser’s e-mail address.