documents: Add support for editing ACLs using Drive v2
A new sub-class was introduced to set JSON as the content-type for
Drive v2. It was a good opportunity to move the Drive-specific JSON
parsing code from the base class.

New API added:
 • GDataDocumentsAccessRule

This new class must be used in place of GDataAccessRule for any
interactions with the Drive API. This is essentially a soft API break,
but it’s unavoidable.

https://bugzilla.gnome.org/show_bug.cgi?id=684920