tests: Fix query parameter comparisons involving dates in YouTube tests
The only way to implement some of the standard feeds in the YouTube
service is now to query filtering against the current date, which means
that a custom message comparison function has to be installed into the
uhttpmock code to ignore that query parameter when replacing HTTP
traces.

Use some new convenience API in libuhttpmock to achieve this.

Dependency changes:
 • libuhttpmock ≥ 0.5.0

https://bugzilla.gnome.org/show_bug.cgi?id=750914