tests: Port and re-enable YouTube video rating tests
A lot of the media:group parsing stuff has gone, and was not tested
properly to begin with, so that test can be dropped. Similarly, the
error handling test can be dropped because the JSON parser code is
deliberately a lot more tolerant now that arbitrary ratings boards are
supported.