documents: Set the content type for GDataDocumentsDocument sub-classes
In Drive v2, every GDataDocumentsDocument sub-class is tied to a
specific content type. Except PDFs, these are the Drive-specific
application/vnd.google-apps.* types. This is unlike the older
Documents v3 API, where you could create a GDataDocumentsDocument to
represent a random ODF on the hard disk, and the server would interpret
it as a GDataDocumentsText instance.

Therefore, we should ensure that instances of these sub-classes have
the corresponding content type attached to them. Otherwise, they will
be identified as application/octet-stream and exportLinks will be
missing.

https://bugzilla.gnome.org/show_bug.cgi?id=684920