tests: ODTs are represented by GDataDocumentsDocument in Drive v2
We are uploading the ODT as a GDataDocumentsDocument, but somehow
expect it to become a GDataDocumentsText when downloading. That's a
bit bizarre. More importantly, in Drive v2, only PDFs and
Drive-specific content types are supposed to be represented by the
GDataDocumentsDocument sub-classes. Therefore, it is wrong to expect
an ODT to somehow become a GDataDocumentsText.

https://bugzilla.gnome.org/show_bug.cgi?id=684920