core: Support pagination using page tokens from JSON
Based on a patch by Milan Crha <mcrha@redhat.com>.

This reworks how pagination is implemented so that multiple pagination
mechanisms are supported explicitly, making the code a lot clearer. A
lot of the new services use pageToken parameters, which we did not
previously support — so this fixes support for pagination in the Google
Tasks service, for example.

This also means that we can drop the hacky pagination support from
GDataDocumentsService.

https://bugzilla.gnome.org/show_bug.cgi?id=775813