calendar: Update query parameters to v3 of the API
Somehow the URI parameters in GDataCalendarQuery did not get updated
with the rest of the code to use the v3 API, meaning that all
non-default query options have been broken since May 2015.

Update them to the v3 parameters, which maintains almost the same
functionality as before. The following properties are no longer
supported in any form on the server, and have been deprecated:

 • GDataCalendarQuery:recurrence-expansion-start
 • GDataCalendarQuery:recurrence-expansion-end
 • GDataCalendarQuery:sort-order