build: Bump libsoup dependency for streams test to 2.55.90
Previously the streams test depended on new API added in libsoup 2.47.3.
It still depends on that API, but it’s become apparent that there was a
bug in libsoup which rendered soup_server_set_ssl_cert_file() useless
until libsoup 2.55.90.

Bump the dependency (for the test code only) to 2.55.90.

Thanks to Mart Raudsepp for their help in debugging.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://bugzilla.gnome.org/show_bug.cgi?id=780081