contacts: Ignore parse errors from websites in contacts
It seems the server permits empty websites (with an empty URI) to be
added to contacts. This seems to contradict the specification (or, at
least, is not mentioned as allowed).

https://developers.google.com/google-apps/contacts/v3/reference#gcWebsite

Rather than accepting such nonsensical websites and exposing them to the
user of libgdata to deal with, it seems better to drop them from the
XML. This means that we don’t round-trip perfectly, but I can’t see how
that can cause problems.

https://bugzilla.gnome.org/show_bug.cgi?id=790671