documents: Support deleting shared Google Drive files
`gdata_documents_service_remove_entry_from_folder` was earlier
performing a PUT request for removing parent folders of a file,
but Drive API disallows (it keeps atleast 1 parents) setting the
"parents" field in PUT request to an empty array. Instead, a
DELETE request to the "parents/delete" needs to be used.

Closes: https://gitlab.gnome.org/GNOME/libgdata/issues/26