meson: Replace custom script with custom targets using sed
Custom scripts are used as a workaround for the namespace issue in
enum sources and headers. However, these scripts instead of using
the `glib-mkenum` program generate both files from scratch.

These avoids the use of these custom scripts to use meson custom
targets along with the use of `sed`. Directories where these headers
are located are also included in the compilation process.