meson: Take advantage of the libgdata internal dependency
`libgdata` internal dependency holds information about the library to
be linked, the location of used headers and the necessary
dependencies. Private dependencies are also not necessary because
they are used only in the compilation process of `libgdata`.

The extra information has been replaced by `libgdata`'s internal
dependency and headers that are also not necessary to be listed have
been removed.