meson: Review libgdata building
The `libgdata` library building, it's GObject Instrospection (GIR),
Typelib and the Vala API file generetion has been reviewed.

This has been done by maximizing the reuse of variables in every
file to avoid the creation of new variables. Four arrays are also
used to hold source file sets for `libgdata`.

The first one the dinamically generated enum headers, that are
publically exposed and must be added to `libgdata` internal
dependency so proper building dependencies are handled.

The second one for the public headers that are used in the GIR and
Typelib data genration. Most of these are also installed in
different directories.

The third one is used for storing the source files located in
different subdirectories. These way these files are decoupled for
the main `gdata` meson build file.

The last one is used to store the different directories holding
header files used in building of the library.