meson: Use libuhttpmock and gdk-pixbuf only on tests
Both `libuhttpmock` and `gdk-pixbuf` are used in the entire project
compilation process. However, they should only be used with tests.

The condition for looking for the `libuhttpmock` is also wrong
because it is needed even if tests are not installed.

This has been changed to use the dependency on the tests compilation
process and removed from the rest.