meson: Fix gtk dependency
`gtk` dependency is a `feature` option with the `auto` value.
However, meson is checking if it's enabled but is not if the value
is `auto`, so the dependency is not checked.

`gtk` dependency is also only used in the `scrapbook`. It has been
removed from `libgdata`'s dependencies and used only when the demo
program is build.

Both issues have been fixed.