docs: Drop use of `<!-- -->` to delimit plurals
gtk-doc gained the ability to magically strip off the ‘s’ in a plural
many years ago, so the workaround of splitting the ‘s’ from a class or
function name is no longer needed.

This clarifies the docs in the code a bit.

This commit was made mechanically using
```
git search-replace -f '<!-- -->s///s'
```

Signed-off-by: Philip Withnall <withnall@endlessm.com>