documents: Added "properties" list to GDataDocumentsEntry
The parse_json and get_json functions have been augmented to support
parsing the "properties" list on a file/folder's metadata. The
"properties" list of GDataDocumentsEntry corresponds to the
"properties" list on a file/folder object's metadata. This list stores
all the public/private custom properties and is writable.

The "properties" list has elements of of type GDataDocumentsProperty,
and functions to search/remove/add such properties have also been added.