documents: Fix gdata_documents_service_add_entry_to_folder() to copy "properties" list too
gdata_documents_service_add_entry_to_folder() copies a given GDataEntry
to a specified parent folder. It did that by creating a local entry.
This local entry is created from the GDataEntry passed as argument, but
it didn't copy over the "properties" list. This commit fixes this
behaviour by copying the "properties" list as well.