build: Make OAuth 1.0 support optional and disabled by default
liboauth is unmaintained and hard to compile at the moment. OAuth 1.0
itself has been deprecated in Google APIs since 2012, in favour of OAuth
2.0 or other more modern authentication/authorisation methods.

Add a new configure option for `oauth1`, and disable it by default. If
it’s disabled, the `GDataOAuth1Authorizer` will return errors when used
(but the API is still available and not marked as deprecated yet).

Eventually the `GDataOAuth1Authorizer` API will be marked as deprecated,
but we can’t do that at the moment as we’re in API freeze.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

Fixes: #1