build: Don’t install tests which are expected to fail
There’s no way to annotate an installed-test to say it’s expected to
fail, so don’t install tests which are expected to fail.

The PicasaWeb tests are expected to fail because the server for them has
been shut down (see #19). The Documents tests are expected to fail
because they haven’t been ported to Drive v2 yet (see #12).

Signed-off-by: Philip Withnall <withnall@endlessm.com>

Fixes: #35