documents: Add Shared Drives support
It seems that the existing documents APIs works nicely with the files
on Shared Drives (formerly Team Drives), although it is possible that
some small changes will be needed to handle differences:
https://developers.google.com/drive/api/v2/shared-drives-diffs

However, the main problem is that there is no way currently to get the
list of Shared Drives IDs and their names. Let's add new APIs for it
which will be consequently used by GVfs.

Relates: https://gitlab.gnome.org/GNOME/gvfs/-/issues/377
Fixes: https://gitlab.gnome.org/GNOME/libgdata/-/issues/25